# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/login": {
            "post": {
                "description": "用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Login.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "login successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_User"
                        }
                    }
                }
            }
        },
...
        "/user/refresh": {
            "get": {
                "description": "刷新登录态",
                "tags": [
                    "User"
                ],
                "summary": "刷新登录态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "刷新用 Token",
                        "name": "X-Refresh-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "nothing",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Register.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "register successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-bool"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.CommonResponse-bool": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "boolean"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-schema_User": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.User"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "schema.Permission": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "是否启用（唯一可设置的字段）",
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "权限描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "module": {
                    "description": "所属模块（handler名称）",
                    "type": "string"
                },
                "name": {
                    "description": "权限名称",
                    "type": "string"
                },
                "path": {
                    "description": "权限路径（形如：POST:/user/create）",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "schema.Role": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "角色描述",
                    "type": "string"
                },
                "display_name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "permissions": {
                    "description": "多对多关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Permission"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "schema.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "description": "用户与角色之间的多对多关系",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Role"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "user.Login.loginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.Register.registerRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}